backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai==1.10.0
anthropic==0.8.1
google-generativeai==0.3.2
sentence-transformers==2.2.2
numpy==1.26.3
scikit-learn==1.4.0
